Mode2&3 structure complete

+++ User/Mode3/Mode3.h
@@ -2,11 +2,14 @@
 // Created by Penta on 2025/7/31.
 //
 
+#include "fastmath.h"
+
+
 #ifndef MODE3_H
 #define MODE3_H
 
 extern uint32_t m3_freq;
-extern uint8_t m3_mag;
+extern double m3_mag;
 
 void Mode3_FreqMagSel(const uint8_t *input);
